git-svn-id: https://oceanai.mit.edu/svn/moos-ivp-aro/trunk@1693 7b6de600-d9ab-45c9-a1ca-e8145d3c6daf

+++ ivp/src/lib_behaviors-marine/BHV_HSLine.h
@@ -34,8 +34,8 @@ public:
   bool         setParam(std::string, std::string);
 
 protected:  // configuration parameters
-  int       m_time_on_leg;
-  double    m_max_des_speed_observed;
+  double    m_time_on_leg;               // seconds
+  double    m_max_des_speed_observed;    // meters/second
 };
 #endif
